+++ app/src/main/java/com/example/togedy_android/presentation/calendar/addPersonalSchedule/AddPersonalScheduleViewModel.kt
@@ -1,11 +1,13 @@
 package com.example.togedy_android.presentation.calendar.addPersonalSchedule
 
+import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class AddPersonalScheduleViewModel @Inject constructor(
 
-) {
+): ViewModel() {
+
 
 }
