+++ app/src/main/java/com/example/togedy_android/presentation/main/MainScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.example.togedy_android.core.navigation.Route
-import com.example.togedy_android.presentation.calendar.navigation.calendarScreen
+import com.example.togedy_android.presentation.calendar.calendar.navigation.calendarScreen
 import com.example.togedy_android.presentation.main.component.MainBottomBars
 import com.example.togedy_android.presentation.community.navigation.communityScreen
 import com.example.togedy_android.presentation.gptscreen.navigation.gptScreen
