+++ app/src/main/java/com/example/togedy_android/presentation/ui/bottom_nav/BottomNavigationBar.kt
@@ -17,7 +17,7 @@
 //import androidx.navigation.compose.currentBackStackEntryAsState
 //import androidx.navigation.compose.rememberNavController
 //import com.example.togedy_android.presentation.ui.model.Routes
-//import com.example.togedy_android.presentation.calendar.CalendarScreen
+//import com.example.togedy_android.presentation.calendar.calendar.CalendarScreen
 //import com.example.togedy_android.presentation.community.CommunityScreen
 //import com.example.togedy_android.presentation.gptscreen.GPTScreen
 //import com.example.togedy_android.presentation.mypage.MyPageScreen
